refactor: Moved Pancake theme out of types to avoid circular dependencies
