+++ src/components/Card/types.ts
@@ -1,3 +1,10 @@
+import { Colors } from "../../theme/types";
+
+export interface CardRibbonProps {
+  variantColor?: keyof Colors;
+  text: string;
+}
+
 export type CardTheme = {
   background: string;
   boxShadow: string;
