+++ src/styled.d.ts
@@ -1,5 +1,5 @@
 import "styled-components";
-import { PancakeTheme } from "./theme/types";
+import { PancakeTheme } from "./theme";
 
 declare module "styled-components" {
   /* eslint-disable @typescript-eslint/no-empty-interface */
