+++ src/theme/types.ts
@@ -1,9 +1,3 @@
-import { ButtonTheme } from "../components/Button/types";
-import { CardTheme } from "../components/Card/types";
-import { TagTheme } from "../components/Tag/types";
-import { ToggleTheme } from "../components/Toggle/types";
-import { NavTheme } from "../widgets/Nav/types";
-
 export type Breakpoints = string[];
 
 export type MediaQueries = {
@@ -54,20 +48,3 @@ export type ZIndices = {
   dropdown: number;
   modal: number;
 };
-
-export interface PancakeTheme {
-  siteWidth: number;
-  isDark: boolean;
-  colors: Colors;
-  button: ButtonTheme;
-  card: CardTheme;
-  tag: TagTheme;
-  nav: NavTheme;
-  toggle: ToggleTheme;
-  breakpoints: Breakpoints;
-  mediaQueries: MediaQueries;
-  spacing: Spacing;
-  shadows: Shadows;
-  radii: Radii;
-  zIndices: ZIndices;
-}
